Worked towards representing all time in Sequentia as an int64_t where 1,000,000 is 1 second.
Added a SeqTime header to express 'SEQ_TIME', implemented some rounding macros there.
Moved TimeBase from Sequentia.h to SeqTime.h.
SeqScene and SeqUISequencer is now only using the new time representation.

+++ sequentia/SeqScene.h
@@ -21,7 +21,7 @@ public:
 	SeqChannel* GetChannelByActionId(const int id);
 	int GetChannelIndexByActionId(const int id);
 
-	double GetLength();
+	int64_t GetLength();
 
 private:
 	int NextActionId();
